requirements.txt
Django==4.1.2
djangorestframework==3.14.0
psycopg2==2.9.3
asgiref==3.5.2
attrs==22.1.0
django-filter==22.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.24.1
inflection==0.5.1
jsonschema==4.15.0
PyJWT==2.4.0
pyrsistent==0.18.1
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
transliterate==1.10.2
django-jazzmin==2.5.0


